+++ docker/examples/Dockerfile
@@ -3,5 +3,5 @@
 FROM ray-project/deploy
 RUN conda install -y -c conda-forge tensorflow
 RUN apt-get install -y zlib1g-dev
-RUN pip install gym[atari] opencv-python==3.2.0.8 smart_open hyperopt
+RUN pip install gym[atari] opencv-python==3.2.0.8 smart_open
 # RUN conda install -y -q pytorch torchvision -c soumith
