+++ test/jenkins_tests/run_multi_node_tests.sh
@@ -214,6 +214,10 @@ docker run --rm --shm-size=10G --memory=10G $DOCKER_SHA \
     python /ray/python/ray/tune/examples/tune_mnist_async_hyperband.py \
     --smoke-test
 
+# docker run --rm --shm-size=10G --memory=10G $DOCKER_SHA \
+#     python /ray/python/ray/tune/examples/hyperopt_example.py \
+#     --smoke-test
+
 docker run --rm --shm-size=10G --memory=10G $DOCKER_SHA \
     python /ray/python/ray/rllib/examples/multiagent_mountaincar.py
